INT-65 Reorganize `Questionnaires`
