+++ src/view/screens/Questionnaires/Inference/index.tsx
